moved mem4 operations to helper module
